Rename insert.py to mutation.py
