+++ pymilvus_orm/partition.py
@@ -15,7 +15,7 @@ import json
 from .exceptions import CollectionNotExistException, PartitionNotExistException, ExceptionsMessage
 from .prepare import Prepare
 from .search import SearchResult
-from .insert import MutationResult
+from .mutation import MutationResult
 from .future import SearchFuture, MutationFuture
